moneymong-209 fix: getErrorByStatusCode 위치 변경
