+++ core/common/src/main/java/com/moneymong/moneymong/common/error/HttpError.kt
@@ -27,3 +27,14 @@ sealed interface HttpError : MoneyMongError {
      */
     data class InternalServerError(override val message: String) : HttpError
 }
+
+fun getErrorByStatusCode(statusCode: Int, message: String): MoneyMongError {
+    return when (statusCode) {
+        400 -> HttpError.BadRequestError(message = message)
+        401 -> HttpError.UnauthorizedError(message = message)
+        403 -> HttpError.ForbiddenError(message = message)
+        404 -> HttpError.NotFoundError(message = message)
+        500 -> HttpError.InternalServerError(message = message)
+        else -> MoneyMongError.UnExpectedError
+    }
+}
